# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Backend API",
        "version": "1.0.0"
    },
    "basePath": "/api/",
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Register user",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User that should be stored",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "password_confirmation": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Login",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "description": "API Login, on success return JWT Auth token",
                                    "type": "string"
                                },
                                "password": {
                                    "description": "API Login, on success return JWT Auth token",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "operationId": "logout",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/recover": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Recover user",
                "operationId": "recover",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Login",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "description": "API Recover Password",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}